+++ tests/QueryType/Update/RequestBuilderTest.php
@@ -93,7 +93,18 @@ class RequestBuilderTest extends TestCase
         );
     }
 
-    public function testBuildAddXmlSpecialCharacters()
+    public function testBuildAddXmlFilterControlCharacters()
+    {
+        $command = new AddCommand();
+        $command->addDocument(new Document(['id' => 1, 'text' => 'test '.chr(15).' 123 '.chr(8).' test']));
+
+        $this->assertSame(
+            '<add><doc><field name="id">1</field><field name="text">test   123   test</field></doc></add>',
+            $this->builder->buildAddXml($command)
+        );
+    }
+
+    public function testBuildAddXmlEscapeCharacters()
     {
         $command = new AddCommand();
         $command->addDocument(new Document(['id' => 1, 'text' => 'test < 123 > test']));
@@ -107,7 +118,7 @@ class RequestBuilderTest extends TestCase
     public function testBuildAddXmlMultivalueField()
     {
         $command = new AddCommand();
-        $command->addDocument(new Document(['id' => [1, 2, 3], 'text' => 'test < 123 > test']));
+        $command->addDocument(new Document(['id' => [1, 2, 3], 'text' => ['test < 123 '.chr(8).' test', 'test '.chr(15).' 123 > test']]));
 
         $this->assertSame(
             '<add>'.
@@ -115,7 +126,8 @@ class RequestBuilderTest extends TestCase
             '<field name="id">1</field>'.
             '<field name="id">2</field>'.
             '<field name="id">3</field>'.
-            '<field name="text">test &lt; 123 &gt; test</field>'.
+            '<field name="text">test &lt; 123   test</field>'.
+            '<field name="text">test   123 &gt; test</field>'.
             '</doc>'.
             '</add>',
             $this->builder->buildAddXml($command)
@@ -136,6 +148,13 @@ class RequestBuilderTest extends TestCase
                                 16,
                             ],
                         ],
+                        [
+                            'nested_id' => 'XLII',
+                            'customer_ids' => [
+                                17,
+                                18,
+                            ],
+                        ],
                         2,
                         'foo',
                     ],
@@ -152,6 +171,11 @@ class RequestBuilderTest extends TestCase
             '<field name="nested_id">42</field>'.
             '<field name="customer_ids">15</field>'.
             '<field name="customer_ids">16</field>'.
+            '</doc>'.
+            '<doc>'.
+            '<field name="nested_id">XLII</field>'.
+            '<field name="customer_ids">17</field>'.
+            '<field name="customer_ids">18</field>'.
             '</doc>'.
             '</field>'.
             '<field name="id">2</field>'.